fix: moved authOptions to utils
