+++ src/components/providers/index.tsx
@@ -4,7 +4,7 @@ import {Toaster} from 'sonner'
 import {ConfigurationProvider} from "@/components/providers/ConfigurationProvider";
 import SessionProvider from './SessionProvider';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/lib/utils';
 
 const AppProviders = async ({children}: { children: React.ReactNode }) => {
     const session = await getServerSession(authOptions)
